better bounds checking for tnef object
This issue was quite interesting as it only appeared when switching from
Go 1.15 to Go 1.16. Naturally the reason is out-of-bounds reading, but
for the specific file we were testing with this wasn't an issue before
in Go 1.15, because the ioutil.ReadFile() method was implemented
slightly different in a way that the difference between len(data) and
cap(data) was bigger and therefore the tnef code didn't panic.